cloud_functions/scrape_trends/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
pytrends==4.*
requests==2.*
